[tflitefile_tool] Separate GraphStats to data and printer
Let's separate GraphStats to data and printer.
